+++ karma.conf.js
@@ -1,11 +1,11 @@
 var webpack = require("webpack");
-var webpackKarma = require("karma-webpack");
 var path = require("path");
+var CircularDependencyPlugin = require('circular-dependency-plugin');
 
 module.exports = function(config) {
   config.set({
     frameworks: ['mocha'],
-    browsers: ['PhantomJS'],
+    browsers: ['PhantomJS', 'Chrome'],
     files: [
       './node_modules/phantomjs-polyfill/bind-polyfill.js',
       'test/index.js',
@@ -59,6 +59,10 @@ module.exports = function(config) {
           ENV_API_ROOT: null,
           ENV_LOGIN_ROOT: null,
           ENV_APP_ROOT: null
+        }),
+        new CircularDependencyPlugin({
+          failOnError: true,
+          exclude: /node_modules/,
         })
       ]
     },
@@ -75,10 +79,11 @@ module.exports = function(config) {
       subdir: '.'
     },
     plugins: [
-      webpackKarma,
+      require("karma-webpack"),
       require("karma-mocha"),
       require("karma-spec-reporter"),
       require("karma-phantomjs-launcher"),
+      require("karma-chrome-launcher"),
       require("karma-sourcemap-loader"),
       require("karma-coverage")
     ],
