+++ package.json
@@ -35,6 +35,7 @@
     "babel-preset-react-hmre": "^1.0.0",
     "babel-root-import": "^4.1.0",
     "chai": "^3.5.0",
+    "circular-dependency-plugin": "^2.0.0",
     "coveralls": "^2.11.9",
     "cross-env": "^1.0.6",
     "enzyme": "^2.3.0",
