+++ src/api/configs/linodes.js
@@ -1,5 +1,7 @@
-import { genConfig, genThunks, genReducer, genActions } from '~/api/gen';
-import { ONE, MANY, PUT, DELETE, POST } from '~/api/gen';
+import {
+  genConfig, genReducer, genActions,
+  ONE, MANY, PUT, DELETE, POST,
+} from '~/api/gen';
 
 export const config = genConfig({
   plural: 'linodes',
@@ -29,5 +31,4 @@ export const config = genConfig({
 });
 
 export const actions = genActions(config);
-export const thunks = genThunks(config, actions);
 export const reducer = genReducer(config);
