+++ src/api/gen.js
@@ -1,4 +1,3 @@
-import { fetch } from '../fetch';
 import _ from 'lodash';
 
 export const ONE = 'ONE';
@@ -13,12 +12,12 @@ export const DELETE = 'DELETE';
 export function genConfig(config, parent = undefined) {
   const result = { ...config, parent };
   if (config.subresources) {
-    Object.keys(config.subresources).forEach(key => {
+    Object.keys(config.subresources).forEach((key) => {
       result.subresources[key] =
-        genConfig(result.subresources[key], result);
+        genConfig(config.subresources[key], result);
     });
   }
-  return result;
+  return Object.freeze(result);
 }
 
 function fullyQualified(resource) {
@@ -50,209 +49,20 @@ export function genActions(config) {
     [MANY]: 'many',
     [DELETE]: 'delete',
   };
-  config.supports.forEach(feature => {
+  config.supports.forEach((feature) => {
     if (typeof actionGenerators[feature] !== 'undefined') {
       actions[fns[feature]] = actionGenerators[feature](config);
     }
   });
   actions.invalidate = () => ({ type: `GEN@${config.plural}/INVALIDATE` });
   if (config.subresources) {
-    Object.keys(config.subresources).forEach(key => {
+    Object.keys(config.subresources).forEach((key) => {
       actions[config.subresources[key].plural] =
-        genActions(config.subresources[key]);
+        genActions(config.subresources[key], 2);
     });
   }
   actions.type = config.plural;
   return actions;
-}
-
-function refineState(config, state, ids) {
-  const path = [];
-  let root = config;
-  const match = (sub, parent) => {
-    if (parent.subresources[sub] === root) {
-      path.push(sub);
-    }
-  };
-  while (root.parent) {
-    const parent = root.parent;
-    Object.keys(parent.subresources).forEach(s => match(s, parent));
-    root = parent;
-  }
-  let refined = state.api[root.plural];
-  const _ids = [...ids];
-  let current = root;
-  let name = null;
-  while (current !== config) {
-    name = path.pop();
-    refined = refined[current.plural][_ids.shift()][name];
-    current = current.subresources[name];
-  }
-  return refined;
-}
-
-function genThunkOne(config, actions) {
-  return (...ids) => async (dispatch, getState) => {
-    let overwrite = false;
-    if (typeof ids[ids.length - 1] === 'boolean') {
-      overwrite = ids.pop();
-    }
-    const state = refineState(config, getState(), ids);
-    const id = ids[ids.length - 1];
-    const prev = state[config.plural][id];
-    if (!overwrite && !_.isUndefined(prev)) {
-      return prev;
-    }
-    const { token } = getState().authentication;
-    const response = await fetch(token, config.endpoint(...ids));
-    const resource = await response.json();
-    dispatch(actions.one(resource, ...ids));
-    return resource;
-  };
-}
-
-function genThunkPage(config, actions) {
-  function fetchPage(page = 0, ...ids) {
-    return async (dispatch, getState) => {
-      const { token } = getState().authentication;
-      const state = refineState(config, getState(), ids);
-      if (state.totalPages !== -1 &&
-          state.pagesFetched.indexOf(page) !== -1) {
-        // cache hit
-        return;
-      }
-      // Update the pages fetched first so we don't double-fetch this resource
-      dispatch(actions.many({
-        page: page + 1,
-        totalPages: -2,
-        totalResults: -2,
-        [config.plural]: [],
-      }, ...ids));
-      const endpoint = `${config.endpoint(...ids, '')}?page=${page + 1}`;
-      const response = await fetch(token, endpoint);
-      const resources = await response.json();
-      actions.many();
-      if (state.totalPages !== -1 &&
-          state.totalResults !== resources.totalResults) {
-        dispatch(actions.invalidate());
-        for (let i = 0; i < state.pagesFetched.length; ++i) {
-          if (state.pagesFetched[i] - 1 !== page) {
-            await dispatch(fetchPage(state.pagesFetched[i] - 1, ...ids));
-          }
-        }
-      }
-      dispatch(actions.many(resources, ...ids));
-      return resources;
-    };
-  }
-  return fetchPage;
-}
-
-function genThunkAll(config, page) {
-  return (...ids) => async (dispatch, getState) => {
-    let state = refineState(config, getState(), ids);
-    if (state.totalPages === -1) {
-      await dispatch(page(0, ...ids));
-      state = refineState(config, getState(), ids);
-    }
-
-    for (let i = 1; i < state.totalPages; i++) {
-      if (state.pagesFetched.indexOf(i + 1) === -1) {
-        await dispatch(page(i, ...ids));
-      }
-    }
-  };
-}
-
-function genThunkUntil(config, actions, one) {
-  return (test, ...ids) => async (dispatch) => {
-    dispatch(actions.one({ _polling: true }, ...ids));
-    for (;;) {
-      try {
-        const resource = await dispatch(one(...ids, true));
-        if (test(resource)) break;
-      } catch (ex) {
-        if (ex.statusCode === 404) {
-          dispatch(actions.delete(...ids));
-          return;
-        }
-        throw ex;
-      }
-      await new Promise(r => setTimeout(r, 3000));
-    }
-    dispatch(actions.one({ _polling: false }, ...ids));
-  };
-}
-
-function genThunkDelete(config, actions) {
-  return (...ids) => async (dispatch, getState) => {
-    const { token } = getState().authentication;
-    const response = await fetch(token, config.endpoint(...ids),
-      { method: 'DELETE' });
-    const json = await response.json();
-    dispatch(actions.delete(...ids));
-    return json;
-  };
-}
-
-function genThunkPut(config, actions) {
-  return (resource, ...ids) => async (dispatch, getState) => {
-    const { token } = getState().authentication;
-    const response = await fetch(token, config.endpoint(...ids), {
-      method: 'PUT',
-      body: JSON.stringify(resource),
-    });
-    const json = await response.json();
-    dispatch(actions.one(json, ...ids));
-    return json;
-  };
-}
-
-function genThunkPost(config, actions) {
-  return (resource, ...ids) => async (dispatch, getState) => {
-    const { token } = getState().authentication;
-    const response = await fetch(token, config.endpoint(...ids, ''), {
-      method: 'POST',
-      body: JSON.stringify(resource),
-    });
-    const json = await response.json();
-    dispatch(actions.one(json, ...ids));
-    return json;
-  };
-}
-
-/**
- * Generates thunks for the provided config.
- */
-export function genThunks(config, actions) {
-  const thunks = { };
-  const supports = a => config.supports.indexOf(a) !== -1;
-  if (supports(ONE)) {
-    thunks.one = genThunkOne(config, actions);
-    thunks.until = genThunkUntil(config, actions, thunks.one);
-  }
-  if (supports(MANY)) {
-    thunks.page = genThunkPage(config, actions);
-    thunks.all = genThunkAll(config, thunks.page);
-  }
-  if (supports(DELETE)) {
-    thunks.delete = genThunkDelete(config, actions);
-  }
-  if (supports(PUT)) {
-    thunks.put = genThunkPut(config, actions);
-  }
-  if (supports(POST)) {
-    thunks.post = genThunkPost(config, actions);
-  }
-  if (config.subresources) {
-    Object.keys(config.subresources).forEach(key => {
-      const subr = config.subresources[key];
-      const plural = subr.plural;
-      thunks[plural] = genThunks(subr, actions[plural]);
-    });
-  }
-  thunks.type = config.plural;
-  return thunks;
 }
 
 function genDefaultState(config) {
@@ -338,7 +148,7 @@ export function genReducer(_config) {
     const { ids } = action;
 
     let name = null;
-    for (let i = 0; i < names.length; i++) {
+    for (let i = 0; i < names.length; i += 1) {
       if (names[i] === config.plural) {
         name = names[i + 1];
         break;
@@ -351,7 +161,7 @@ export function genReducer(_config) {
     const keys = Object.keys(config.subresources);
     let subkey = null;
     let subconfig = null;
-    for (let i = 0; i < keys.length; i++) {
+    for (let i = 0; i < keys.length; i += 1) {
       subkey = keys[i];
       subconfig = config.subresources[subkey];
       if (subconfig.plural === name) {
