+++ src/api/index.js
@@ -1,5 +1,13 @@
-export { thunks as distros } from './configs/distributions';
-export { thunks as datacenters } from './configs/datacenters';
-export { thunks as types } from './configs/types';
-export { thunks as linodes } from './configs/linodes';
-export { thunks as kernels } from './configs/kernels';
+import * as distributionsModule from './configs/distributions';
+import * as datacentersModule from './configs/datacenters';
+import * as typesModule from './configs/types';
+import * as linodesModule from './configs/linodes';
+import * as kernelsModule from './configs/kernels';
+
+import genThunks from './genThunks';
+
+export const distributions = genThunks(distributionsModule.config, distributionsModule.actions);
+export const datacenters = genThunks(datacentersModule.config, datacentersModule.actions);
+export const types = genThunks(typesModule.config, typesModule.actions);
+export const linodes = genThunks(linodesModule.config, linodesModule.actions);
+export const kernels = genThunks(kernelsModule.config, kernelsModule.actions);
