+++ src/fetch.js
@@ -1,8 +1,9 @@
-import { API_ROOT } from './constants';
 import * as isomorphicFetch from 'isomorphic-fetch';
+
 import store from '~/store';
 import checkLogin from '~/session';
 import { setToken } from '~/actions/authentication';
+import { API_ROOT } from './constants';
 
 /*
  * Sinon cannot stub out a function in a function-only module.
@@ -38,7 +39,7 @@ export function fetch(token, _path, _options) {
   const path = API_ROOT + _path;
   const promise = fetchRef(path, options);
   return new Promise((accept, reject) => {
-    promise.then(response => {
+    promise.then((response) => {
       const _status = response.headers.get('X-Status');
       const status = _status ? parseInt(_status, 10) : response.status;
       // eslint-disable-next-line no-param-reassign
